+++ tests/pmarc2.rs
@@ -6,19 +6,19 @@ mod sink;
 use sink::SinkSum;
 use CompressionMethod::*;
 
-const TESTS_CASES: &[(u64, &str, &str, u64, u64, u16, u32, OsType, CompressionMethod)] = &[
-    (0, "pm0.pma",     "GPL-2.GZ",    6912,    6912, 0x39CC, 0x549D935A, OsType::Generic, Pm0),
-    (0, "pm2.pma",     "GPL-2.",      7098,   18176, 0x83CD, 0x8E2093A7, OsType::Generic, Pm2),
-    (0, "long.pma",    "LONG.TXT",   85397, 1241659, 0x2AEA, 0xB2B419D6, OsType::Generic, Pm2),
-    (0, "comment.pma", "HELLO.TXT",     22,     128, 0x9784, 0x906330C5, OsType::TownsOs, Pm2),
-    (0x86D, "sfx.com", "GPL-2.",      7098,   18176, 0x83cd, 0x8E2093A7, OsType::Generic, Pm2),
+const TESTS_CASES: &[(u64, &str, &str, u64, u64, u16, u32, OsType, CompressionMethod, bool)] = &[
+    (0, "pm0.pma",     "GPL-2.GZ",    6912,    6912, 0x39CC, 0x549D935A, OsType::Generic, Pm0, false),
+    (0, "pm2.pma",     "GPL-2.",      7098,   18176, 0x83CD, 0x8E2093A7, OsType::Generic, Pm2, false),
+    (0, "long.pma",    "LONG.TXT",   85397, 1241659, 0x2AEA, 0xB2B419D6, OsType::Generic, Pm2, false),
+    (0, "comment.pma", "HELLO.TXT",     22,     128, 0x9784, 0x906330C5, OsType::Generic, Pm2, true),
+    (0x86D, "sfx.com", "GPL-2.",      7098,   18176, 0x83cd, 0x8E2093A7, OsType::Generic, Pm2, false),
 ];
 
-const EXTENDED_COMMENT: &str = "his is a comment attached to the file hello.txt.";
+const EXTENDED_COMMENT: &str = "This is a comment attached to the file hello.txt.";
 
 #[test]
 fn test_pmarc2() -> io::Result<()> {
-    for (offset, name, path, size_c, size_o, crc16, crc32, ostype, compr) in TESTS_CASES {
+    for (offset, name, path, size_c, size_o, crc16, crc32, ostype, compr, has_comment) in TESTS_CASES {
         println!("-------------\n{:?}", name);
         let mut file = fs::File::open(format!("tests/pmarc2/{}", name))?;
         file.seek(SeekFrom::Start(*offset))?;
@@ -26,8 +26,12 @@ fn test_pmarc2() -> io::Result<()> {
         loop {
             let mut sink = SinkSum::new();
             let header = lha_reader.header();
-            if !header.extended_area.is_empty() {
+            if *has_comment {
+                assert_eq!(header.parse_comment().unwrap(), EXTENDED_COMMENT);
                 assert_eq!(str::from_utf8(&header.extended_area).unwrap(), EXTENDED_COMMENT);
+            }
+            else {
+                assert!(header.parse_comment().is_none());
             }
             assert_eq!(header.level, 0);
             assert_eq!(header.msdos_attrs, MsDosAttrs::ARCHIVE);